requirements.txt
pytesseract>=0.3.10
Pillow>=9.0.0
opencv-python-headless>=4.5.0
edge-tts>=6.1.0